Medical app backed stage02

+++ medical-app-backend/src/main/java/com/fral/medapp/controller/DoctorRestController.java
@@ -1,7 +1,10 @@
 package com.fral.medapp.controller;
 
+import java.net.URI;
 import java.util.ArrayList;
 import java.util.List;
+
+import javax.validation.Valid;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
@@ -15,7 +18,9 @@ import org.springframework.web.bind.annotation.PutMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
+import org.springframework.web.servlet.support.ServletUriComponentsBuilder;
 
+import com.fral.medapp.exception.EntityNotFoundException;
 import com.fral.medapp.model.Doctor;
 import com.fral.medapp.service.DoctorService;
 
@@ -35,47 +40,40 @@ public class DoctorRestController {
 	
 	@GetMapping(value = "/{id}", produces = MediaType.APPLICATION_JSON_VALUE)
 	public ResponseEntity<Doctor> listarId(@PathVariable("id") Integer id) {
-		Doctor medico = new Doctor();
-		medico = service.findById(id);
+		Doctor medico = service.findById(id);
+		
+		if (medico == null) {
+			throw new EntityNotFoundException("Id: " + id);
+		}
 		return new ResponseEntity<Doctor>(medico, HttpStatus.OK);
 	}
 
 	@PostMapping(consumes = MediaType.APPLICATION_JSON_VALUE, produces = MediaType.APPLICATION_JSON_VALUE)
-	public ResponseEntity<Integer> registrar(@RequestBody Doctor medico) {
-		int resultado = 0;
-		try {
-			service.create(medico);
-			resultado = 1;
-		} catch (Exception e) {
-			resultado = 0;
-		}
+	public ResponseEntity<Object> registrar(@Valid @RequestBody Doctor medico) {
+		
+		service.create(medico);
+		URI location = ServletUriComponentsBuilder.fromCurrentRequest().path("/{id}")
+					   .buildAndExpand(medico.getId()).toUri();
 
-		return new ResponseEntity<Integer>(resultado, HttpStatus.OK);
+		return ResponseEntity.created(location).build();
 	}
 
 	@PutMapping(consumes = MediaType.APPLICATION_JSON_VALUE, produces = MediaType.APPLICATION_JSON_VALUE)
-	public ResponseEntity<Integer> actualizar(@RequestBody Doctor medico) {
-		int resultado = 0;
-		try {
-			service.update(medico);
-			resultado = 1;
-		} catch (Exception e) {
-			resultado = 0;
-		}
-
-		return new ResponseEntity<Integer>(resultado, HttpStatus.OK);
+	public ResponseEntity<Object> actualizar(@Valid @RequestBody Doctor medico) {
+		
+		service.update(medico);
+		return new ResponseEntity<Object>(HttpStatus.OK);
 	}
 
 	@DeleteMapping(value = "/{id}", produces = MediaType.APPLICATION_JSON_VALUE)
-	public ResponseEntity<Integer> eliminar(@PathVariable("id") Integer id) {
-		int resultado = 0;
-		try {
-			service.delete(id);
-			resultado = 1;
-		} catch (Exception e) {
-			resultado = 0;
+	public void eliminar(@PathVariable("id") Integer id) {
+		
+		Doctor doctor = service.findById(id);
+		if (doctor == null) {
+			throw new EntityNotFoundException("Id: " + id);
 		}
+		
+		service.delete(id);
 
-		return new ResponseEntity<Integer>(resultado, HttpStatus.OK);
 	}
 }
